requirements.txt
django>=4.2,<6.1
pdfminer.six>=20221105
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
